Moved beam-server to package
